mypy --strict comes back clean over 104 sources of the app package, and the
checker is declared and locked so it is not a tool that only exists on the
machine of whoever ran it. The lint gate does not call it yet: that is phase 2,
with the other two packages clean, because a gate born red is a gate someone
switches off.

The blind validator served the real base against both revisions and compared the
HTML of the four money screens byte for byte. They are identical. Two type:
ignore survive, both on the same conflict between the Starlette stub and the
cookie casing a test pins.

+++ app/design/contrast.py
@@ -11,7 +11,7 @@ def _channel(value: int) -> float:
     share = value / 255
     if share <= 0.03928:
         return share / 12.92
-    return ((share + 0.055) / 1.055) ** 2.4
+    return float(((share + 0.055) / 1.055) ** 2.4)
 
 
 def _relative_luminance(color: str) -> float:
